requirements.txt
# Used in non test files
assemblyline
python-baseconv
elastic-apm[flask]

# Only used in test files
fakeredis[lua]